+++ internal/delivery/http/request/user_request.go
@@ -24,6 +24,8 @@ type UserCreateRequest struct {
 	Department        *string           `json:"department,omitempty" binding:"omitempty,max=100"`
 	BankAccountNumber *string           `json:"bank_account_number,omitempty" binding:"omitempty,max=100"`
 	BasicSalary       *float64          `json:"basic_salary,omitempty" binding:"omitempty"`
+	PositionAllowance *float64          `json:"position_allowance,omitempty" binding:"omitempty"`
+	OtherAllowance    *float64          `json:"other_allowance,omitempty" binding:"omitempty"`
 	ProfilePictureID  *string           `json:"profile_picture_id,omitempty" binding:"omitempty,uuid"`
 }
 
@@ -56,6 +58,8 @@ type UserUpdateRequest struct {
 	Department        *string           `json:"department,omitempty" binding:"omitempty,max=100"`
 	BankAccountNumber *string           `json:"bank_account_number,omitempty" binding:"omitempty,max=100"`
 	BasicSalary       *float64          `json:"basic_salary,omitempty" binding:"omitempty"`
+	PositionAllowance *float64          `json:"position_allowance,omitempty" binding:"omitempty"`
+	OtherAllowance    *float64          `json:"other_allowance,omitempty" binding:"omitempty"`
 	ProfilePictureID  *string           `json:"profile_picture_id,omitempty" binding:"omitempty,uuid"`
 }
 
@@ -88,6 +92,8 @@ type UserUpdateProfileRequest struct {
 	Position          *string           `json:"position,omitempty" binding:"omitempty,max=100"`
 	Department        *string           `json:"department,omitempty" binding:"omitempty,max=100"`
 	BankAccountNumber *string           `json:"bank_account_number,omitempty" binding:"omitempty,max=100"`
+	PositionAllowance *float64          `json:"position_allowance,omitempty" binding:"omitempty"`
+	OtherAllowance    *float64          `json:"other_allowance,omitempty" binding:"omitempty"`
 	ProfilePictureID  *string           `json:"profile_picture_id,omitempty" binding:"omitempty,uuid"`
 }