+++ migrations/20260301153134_create_fk_users_files.down.sql
@@ -1,2 +1,0 @@
-ALTER TABLE users
-DROP FOREIGN KEY fk_users_profile_picture_id;