+++ migrations/20260301153134_create_fk_users_files.up.sql
@@ -1,5 +1,0 @@
-ALTER TABLE users
-    ADD CONSTRAINT fk_users_profile_picture_id
-    FOREIGN KEY (profile_picture_id) REFERENCES files(id)
-    ON DELETE SET NULL
-    ON UPDATE CASCADE;